Serialize, deserialize and persist frame tracks between captures
Frame tracks are now serialized when saving captures and are properly
deserialized when loading a capture. The user choice of a frame track
is also persisted between captures, that is, when a user selects a
function as a frame tracks, and takes a new capture, the frame track
is properly added at the end of the capture (what is still missing
are live updates to the frame track during capture).

Bug: http://b/171026228
Tested: Manual tests (capture/save/load, etc.)