Initialize EventLoop in the background thread
Previously ServiceDeployManager's additional event loop object has been
initialized on the main thread and then called in the background thread.

This seems not to work under Linux. Initializing the event loop from the
thread where it's going to be called works.

It's still unclear whether that's a bug or a documentation error.