Live updates to frame tracks during a capture
We now update frame tracks live during a capture by processing each
timer that belongs to a function call, creating a frame track timer
from its timestamp and the previously encountered timestamp for this
function call.

Note that timers may arrive out of order, in particular when a frame
track function was chosen that is called from multiple threads. During
live capture we do not correct for this, but at the end of the capture,
the entire frame track is refreshed to correct any mistakes made during
capture.

Bug: http://b/173100696
Tested: Unit tests; various manual tests, including serialization of
frame tracks.