Don't clamp process cpu usage to 100%
A process's cpu usage is considered 100% when it fully occupies a single
logical core. This means a process can easily reach more than 100% cpu
usage by performing tasks on different cores in parallel.

This was not properly handled in the new process utility code which used
to clamp to 100%. This commit fixes that.

Test: Manual test with Trata
Bug: http://b/172332478