Move the view when dragging a track to the border
Before this, we couldn't move a track behind or above to some track that
wasn't in the screen. Now if we drag a track to the top or to the
bottom, the view is going to move.

Also, we refactored and renamed UpdateTopLeftY so each time that we update
where is the topY on the screen, we check that fits in the min-max range
of the slider.