Remove ServiceDeployManager::loop_
loop_ used to be a member variable of ServiceDeployManager which has
been used in a couple of function rather independently.

The recent threading changes have made the usage of loop_ even more
brittle because due to a bug in Qt it has to be constructed on the
thread that will eventually use the event loop.

This highlights that we use QEventLoop/EventLoop in a way it was not
designed for. We should instantiate it locally whenever we need it.

So this commit replaces the "global" long-living loop_ by local
short-living instances inside of the member functions.

That makes the code easier to understand and we don't need to rely on
comments to protect the previous rather brittle behaviour.