Use std::optional for CaptureData to indicate when it is available
We have relied on CaptureData to be empty in a sane way when no actual
capture data is available, for example, we draw empty captures and we
save empty captures, even when no capture was loaded or captured.

We now use std::optional to indicate whether actual capture data is
available and react accordingly when drawing, updating, and saving
captures. That is, in particular when no capture data is available,
these actions are skipped or not allowed (users now cannot save anymore
when no capture is available).

This is mainly a refactor, but some user visible changes were needed to
disallow saving captures when no capture is available.

Tested: Ran unit tests; lots of manual testing with loading/saving/
capturing profiles, etc.