Move UserDefinedCaptureData from CaptureData to DataManager only
The recent refactoring of CaptureData to using std::optional and the
need of adding frame tracks before a capture makes it clear that
UserDefinedCaptureData (which owns frame track data) should not be a
member of CaptureData (and indeed does not have to be a member).

This is a mostly mechanical change that removes the
UserDefinedCaptureData member from CaptureData.

Tested: Ran unit tests; manual load/save/capture tests with and without
presets.

Bug: b/173191435