Add option for enabling a frame track for a function to Functions tab
So far it was only possible to add frame tracks after taking a capture.
With this change, it is now possible to enable a frame track (or
multiple) even before the first capture. This is possible from the
Functions tab by choosing the context menu option
"Enable frame track(s)".

Note that a function that is chosen as a frame track must be hooked, so
if a function is not hooked when enabling it as a frame track, it is
automatically hooked.

Bug: http://b/173191435

Tested: Lots of manual tests with taking captures, loading/saving
presets, loading/saving captures.