Set parent properly in orbit_qt::EventLoop
This sets the parent property for all member objects
in orbit_qt::EventLoop. The Qt-parent-child system helps
with managing object lifetimes (not needed in this case)
and thread associativity. The latter is important here.
Whenever the EventLoop object is moved to a different thread the child
(QEventLoop) needs to be moved as well.