Add OverlayWidget class
This class is an overlay above any other large enough widget
(e.g. a table). The overlay provides an animation to signal to the user
that something is happening an the program is not frozen. It also provides
a status message field and a button to cancel. Whoever uses the
OverlayWidget has to make sure to react to the "Cancelled" signal, and
shutdown the overlay afterwards (either setVisible(false) or remove all
together)