Add Grpc Protos used for the Vulkan Layer.
This adds a CaptureEvent proto for a Vulkan submission, which gathers
information from a single submission (VkQueueSubmit). In particular
it contains the timings for all the command buffers in that submission
and the debug markers opend/closed in that submission.

These protos can be then used by the layer to send information to the
service.

Test: Compile