Fix issues with frame tracks when loading a capture
When loading a capture, Orbit can be in a state where hooking a function
that has hits in the "Live" tab must not be allowed, either because
Orbit is not in a connected state (no process selected) or because the
module to which the function belongs is not loaded. In both these cases,
the "Hook" option is disabled in the "Live" context menu, however it was
incorrectly possible to hook a function by enabling a frame track. In
the connected state, this leads to a crash when taking a capture with
the incorrectly hooked function.

We now apply the same check that is used for enabling the "Hook" context
menu entry when enabling the frame track for a function to make sure
it's only hooked when possible.

Bug: http://b/174296341

Tested: Manual tests for the cases described on the bug.