Allow Windows builds with system-installed Qt
We can't provide binary packages of Qt for the public and compiling
Qt from source takes a long time and is cumbersome.

To solve that problem this introduces the possibility
to use a system-installed Qt as we support it on Linux
for a long time.

This new support is optional and we will continue building
our own Qt for the official Stadia version of Orbit. But
for external contributors that should simplify getting started
on Windows.