Client side introspection UI
Activate client side introspection. In "--devmode", a new button in the
capture view toolbar appears. Pressing it will instantiate an always-on-top
IntrospectionWindow that can be used to profile Orbit inside of Orbit.
To start and stop a capture, press 'X' just like in the normal capture window.
When the IntrospectionWindow is closed, recording automatically stops.
You can add profiling markers in the Orbit code base by using the manual
instrumentation API in Orbit.h.