Move Classes from OrbitCore to OrbitClientData and OrbitClientModel
This is another step of getting rid of OrbitCore. It moves all the
data classes to OrbitClientData and basically SamplingProfiler and
CaptureData to OrbitClientModul.
After http://b/166238019 is fixed, we can also move CaptureData to
OrbitClientData, as it will be then not dependent to SamplingProfiler
anymore (but only the resuling data class).

TracepointEventProcessor had a dependency to Threading.h, that should
be not moved. So in order to move that class, this also uses
absl::Mutex here, instead of using the typdefs from Threading.h.

Test: Compile & run tests