Add CMake target include-what-you-use
This adds a new CMake target "iwyu" which is not included in the default
ALL target. This target will call include-what-you-use. All code
generation steps are dependencies of this target, so it can operate on a
clean build directory and will do the necessary things.

The results is a iwyu_results.txt file in the build directory.

To run this you need to have `iwyu` installed and it only works on
Linux.

This commit also provides integration into conan. There is a new conan
profile called `iwyu`. You can build that profile to run
include-what-you-use.