Add presubmit check for include-what-you-use
This adds a kokoro presubmit check which is based on the recent iwyu
addition.

This presubmit check is handled as a generic build in the first place
since iwyu needs the whole toolchain set up. We "compile" the conan
profile "iwyu" which generates a results file.

This also adds an additional section to the generic build.sh script that
analyses the results file and prints the issues to the build-log.