Create Track Manager class
We are creating this class which will be on charge of everything related
to the tracks (creation, searching, erasing and sorting). We want to make
unit tests on the tracks and also TimeGraph is currently very big.
For now, this class have a reference to TimeGraph but in the future we
want to erase it.