Fix crash on capture abort on exit
Change `TryAbort` to `AbortCaptureAndWait` and wait (for a limited time) for the
`CaptureClient::Capture` thread to change `state_` to `State::kStopped`, so that
`client_context_` has been cleaned up.
This is mostly done by improving synchronization in the class, adding a mutex
for `client_context_` and `reader_writer_`.

Bug: http://b/175211279

Test: locally and with Trata both taking "normal" captures and aborting captures
on exit.