Fix Vulkan validation layer include path
Depending on the version of the vulkan validationlayers header package,
the include path is different. Old versions put the headers into the
vulkan/ subdirectory, laters ones don't use that sub-directory.

Since we have to work with both versions, this commit adds a cmake
find-package detection script which adds a cmake target
Vulkan::ValidationLayers which conveys for the proper include path.

The vk_layer_dispatch_table.h header is not self-contained and needs a
special inclusion order. That's why this commit als adds some
clang-format descriptors which avoid reordering of the vulkan layers.
Ideally this will be fixed upstream.