Instrumentation to monitor number of  capture data bytes sent
Add a small amount of instrumentation in OrbitService to monitor the
number of bytes of capture data sent during a capture. The average
number of bytes per CaptureEvent message sent is already recorded for
every batch of events sent and shown as a live graph when --devmode
is enabled.

Tested: Manual test.
Bug: http://b/175383580