Capture Window accessibility basics.
Introduces the accessibility interface to be used for controls beneath
the OrbitGlWidget, and unit tests where possible. The interface is
implemented for GlCanvas in this PR, but returns 0 valid children.

Functionality for GlCanvas, TimeGraph and Tracks will follow in a
subsequent PR.

For documentation, check OrbitQt/AccessibilityAdapter.cpp (yes, in the
cpp, because you're expected to not be exposed to this unless extending
the core functionality).