Add Vulkan Layer Controller
The `VulkanLayerController` glues together all the logic of the
Vulkan layer by having an representative for every instrumented
Vulkan function ("OnFunctionName").

It is also responsible for the bootstrapping (in particular
vkCreateInstance/Device) and the enumeration functions to expose
the layer's properties.

Bug: http://b/168200502
Test: Compile & Run tests

TODO: The commit introducing the grpc channel is missing, thus
the tests will fail, at the moment.