Add OrbitProducerSideChannel module defining Unix socket for producers
This defines the Unix socket over which the communication between producers of
`CaptureEvent`s and `OrbitService` will happen.
In the near future, it will be used by `OrbitVulkanLayer` and `OrbitService`.
The name is what it is but it is for consistency with "producer side" used
elsewhere to define the side of `OrbitService` that will communicate with
producers.

Bug: http://b/174028631