Free widget resource in OrbitMainWindow's destructor
Widgets in OrbitMainWindow are destructed AFTER OrbitMainWindow is
destructed, due to Qt rules. This is problematic when these widgets
require resources that live in OrbitMainWindow as these resources are
destructed with OrbitMainWindow - leaving the widgets with dangling
pointers.

This commit adds calls to the widget's deinitialize methods which
release all required resources that are not managed by Qt anyway.