Add static_assert to ensure floating point adheres to IEEE 754 (IEC 559)
We have a couple of places were it's more convenient to possibly divide
by 0.f or 0.0 instead of checking. This is fine as long as floating
point computations adhere to IEEE 754, o/w it's undefined behavior. We
add static_assert's to check for this. The asserts are added in all
(known) places where we may divide by 0.0 or 0.f to ensure we still have
the assert when some of the places are changed or removed.

Test: Compiled.
Bug: N/A