Clean TimeGraph dependences from TrackManager/Track
In the process of making UnitTests for the Tracks, we want to erase
most dependences to TimeGraph to easily test TrackManager. Also we want
to get rid of the circular dependence from TimeGraph and the Tracks.

In this PR we:

 - Use correctly TimeGraph static methods.

 - Get CaptureData from App instead of TimeGraph (CaptureData now isn't
 on TimeGraph anymore).

 - Get StringManager from App instead of passing throught the classes.
 It is only used in GpuTrack.

 - Move initialization (Name, Label, ...) inside the Tracks and no inside
 TrackManager. Now GetOrCreate methods are really similar and could be
 templatized in the future or we can have an only list of tracks and
 easily get from there.

Bugs related:
https://b/176086740
https://b/176056427

Tested with the new_ui: Start a new capture/Load a capture/End session
and load it again.