Keep user selection in sampling tab when filtering
With this change, the sampling view can keep selections made by users
when doing filtering and sorting.
When user edits the filtering conditions on the sampling view, both the
callstack report and the highlighted functions in the callstack report
will be updated according to the visible selected functions.
When user removes the filtering conditions, the sampling view will show
previous selections and update the callstack report accordingly.