Fix the selection behavior in live tab
With this change, the Live tab can keep the selection made by users
when doing filtering and sorting.
When the user edits the filtering condition in the Live tab, the
highlighted functions in both the capture window and Live tab will be
updated according to the visble selected index.
When the user removes the filtering condition, the highlighted
functions in both the capture window and the Live tab will be updated
according to user's previous selection.