OrbitBase pragma for Orbit.h
Orbit.h currently lives in the top-level-directory - which is not
covered by any include directory. That means it needs to be include with
a relative path using `../`. That's a hack and include-what-you-use does
not like that. An IWYU pragma silences the error. This won't be needed
anymore when Orbit.h moves into its own cmake target and subdirectory.