Add versioning support to docker containers
So far we've always been using the `latest` tag for docker containers
which means the CI has always been using a single version of a docker
container.

This change introduces container versions which allow to make breaking
changes to docker containers that will not affect older builds or other
branches.

Instead of always using the `latest` tag, the current docker tag
revision will be stored in a file. The developer has to update this file
when they make a (breaking) change to a docker container.