Fix clang_format/license_header checks on release branches
These checks were broken due to the recently introduced automatic
detection of git target branch names. The kokoro variable contains the
branch name but for a checkout we also need to specify the remote name
`origin`. It happened to work for `master` because it's the default
branch and a local copy exists. That's not the case for a release
branch.

Bug: http://b/175678721
Test: Tested by creating a PR against an old release branch