Use `switch` instead of `if(...) else if(...)` in TimerGraph.
This changes the switch over the different timer types from
an `if(...) else if(...)` construct to a switch statement,
which makes the code more readable and error prone (we will fail
if we miss a case).
Also it avoids having a "default" case (as present before) to
make the code more robust by explicitly stating the expected
case.

Test: Compile