Draw timebar lines again
Regression from 1.55 (new drawing order).
We are using a wrong coordinate system to draw time bar lines.
Before 1.55 we had two different batchers to draw items for each
coordinate system and we had merged both in one.

In this PR:

 - We draw time bar lines using screen space coordinate system.

 - We make a function to convert Qt-screen to Gl-screen

 - We added a comment after a discussion with Pierric about Text should
 be drawn in screen space.
So, we added an if-condition to draw timebar lines in WorldSpace
system coordinates.

Also, we added a comment after a discussion with Pierric to say Text
should be drawn in ScreenSpace coordinates.

Related to http://b/175532525.

Test: Run Orbit + Load a Capture.