Indicate an ended target process in new beta UI
Since the process list was removed in the new (beta) UI it is
particurlary hard to realize that the target process has ended on the
instance.

This commit is adding a somewhat hacky fix. It's checking the regularly
incoming process list for existence of the target process. If the target
process pid cannot be found the corner widget will color the process
name in orange - indicating that the process has died. Additionally it
will disable the Start-Capture button.

Bug: http://b/176960277