Add GetFunctionBaseAddressByAddress to CaptureData.
This adds a `GetFunctionBaseAddressByAddress` to `CaptureData` that
computes the functions's base address from an absolute address
inside a function based on either `ModuleData` or `LinuxAddressInfo`.
If both information are not available, it assumes the give address
is the base address.

This information was already computed internally in the
`SamplingDataPostProcessor`, but not made available. Instead of
just making this available in the `PostProcessedSamplingData`,
the new function goes to `CaptureData`, where sematically close
functions like `GetFunctionNameByAddress` are located.
Inside the `SamplingDataPostProcessor`, we now use the new function,
but still store the results in the map to avoid recomputations.

Bug: http://b/176970009
Test: Take a capture validate callstacks, load module and redo
      the validation.