Handle Vulkan-based Timers in GpuTrack.
This adds support for Vulkan debug marker and command buffer timers
to the GpuTrack.
Command buffer timers will be drawn underneath their corresponding
hw execution timer, which requires to double the depth in the case
that we do have those timers.

For now, debug markers will be drawn on a separate track (named by
the timeline + "_marker"). For the future, we want to combine those
into an composed track.
Colors of the debug markers will either be picked up by the user's
choice, or be determined by the marker's text string.

Bug: http://b/176809754
Test: Compile -- Not really testable without producing the timers,
      which will be the next PR.