Add/Remove FrameTracks properly while capturing
Frame tracks are a special track because could be added or erased while
capturing. In this PR we are solving a few issues about inconsistent
information when adding or removing a frame track while capturing.
Basically, we updated:

 - FrameTrackOnlineProcessor in App (info while capturing)

 - Visible Tracks in TrackManager after removing Frame Tracks.

There are still an issue about multithreading problems (more information
in http://b/176959931).

Test: Start a capture, enable/disable Frame Tracks/ Stop a capture
several times.