Prevent Vulkan queue submissions from being sent out of order.
On every vkQueuePresentKHR, we try to retrieve timestamp results
for all unprocessed submissions. It may happen that on one `present`
some submissions are not yet ready, and on the next `present` we
have multiple ones to process.

Before this change, this could lead to us sending submissions of
a single queue out of order, as we might have picked the latest
submission first to check.

It would complicate the client's processing code, if it is not
guaranteed that submissions of a single queue are retrieved in
order.

Thus, this sorts the submissions by their CPU timestamps after
they are retrieved.

Test: Run the new test