Small fix for frame track UI
After taking a capture and when unhooking a function that was enabled as
a frame track, the frame track is now removed and the frame track icon
is not shown (was shown before). This is more consistent behaviour than
before where it was possible to end up in a state where the frame track
was enabled for the current capture but disabled for the next capture
(with confusing UI options in the context menu).

Bug: http://b/177302941
Tested: Manual tests (take capture, unhook, load capture with frame
track).