Handle Vulkan events in CaptureEventProcessor.
This adds support of the Vulkan related capture events (GpuQueueSubmission)
to CaptureEventProcessor.
For each command buffer and completed debug marker of a queue submission,
we create a TimerInfo object.

In order to convert the GPU timestamps to CPU timestamps, we try match each
queue submission to the corresponding tracepoint-based GpuJob. This requires
us to store the GpuJob/GpuSubmission until the counter-part has been processed
(remember, that the events origin from different data collectors).

If debug markers span across different submissions and we haven't captured
the "begin" marker, we do not discard this timer, but instead use the
begin of capture as start date.

Test: Run tests.
Bug: http://b/176809754.