Use (duration,end) instead of (begin,end) for CaptureEvent gRPC protos
This saves some bandwidth thanks to variable-length encoding of protobufs.
More details in the bug.
GPU-related events are (at least for now) not included in the change. As the
frequency of those events is relatively low compared to scheduler, thread state,
dynamic instrumentation, that would in any case have a smalelr impact.
This does not affect "client protos" (i.e., serialization to file), so that we
don't break compatibility with previous versions.

Bug: http://b/177222925

Test: Update unit tests. Take a few captures on Trata and Infiltrator and verify
that things look good.