Exclude renamed files from presubmit checks
Both clang_format- and license_headers-checks rely on `git diff` to
figure out what has been changed in the PR. This commit adds an option
to the git-diff calls which makes it filter-out renamed files.

So the presubmit checks won't fail if a file was renamed, but not
modified. This policy is in line with what the clang-format check does
anyway by restricting to modified lines.