Optimize EventTrack: don't iterate over all events in CallstackData
Only iterate on the CallstackEvents in the visible range.
We already had `std::map`s in place to do this, I guess something was lost
during some old refactoring.
Of course this doesn't make a difference when completely zoomed out, but
otherwise it makes the capture windows so much smoother with long captures.

Bug: http://b/177523046