Fix crash when enabling frame tracks for manual instrumentation fcts
Functions dynamically instrumented for manual instrumentation (that is,
functions with OrbitType not kNone) show up in the "Live" functions
view (which needs to be changed at some point) and it is possible to
enable a frame track for them. Due to the case distinction in
TimeGraph::ProcessTimer that checks for those special functions with
OrbitType != kNone, when adding a frame track, frame track timers were
processed incorrectly. This is fixed here.

Tested: Manual test with OrbitTest
Bug: http://b/177432700