Draw frametrack text in the correct layer
We were drawing Frame-track-text as text on top in the capture view. So,
this text was drawn on top of other pinned/moving tracks and also over
the green overlay and iterators.

This PR erased completely the text z-layer because it makes confusion
and draw some text in the new track-text layer and other in UI-text
layer.

It solved http://b/177420768.

Test: Load a capture, enable a frame track, add iterators/overlays and
move tracks one on top of another several times.