AccessibilityAdapter instances are now managed by QT
This is the main change to get rid of the exception on shutdown:
Instances of AccessibilityAdapter are no longer managed by
AdapterRegistry, but created through a QT Accessibility Factory and
correctly destroyed when the corresponding QObject is destroyed.

For this to work, a call to InstallAccessibilityFactories() is now
mandatory before creation of any objects, even in the unit tests.