Hide manual instrumentation API functions in live view. (#1673)
The internal API functions for manual instrumentation show up in
the live functions view. This is missleading, as they are not
the results of manual instrumentation.

Also related: Adding a frame track for those is currently leading
to a crash. Hiding those, would also prevent the user for adding
a frame track (at least from this view).

Test: Capture OrbitTest.
Bug: http://b/177432700.