Copy the Vulkan layer to the build artifacts.
This change prepares the deployment of the Vulkan layer.
When building the Vulkan layer, we now copy the .json file
to the lib directory.
In the kokoro build script, we now also copy the layer's
shared object file and the json file, such that it will be
part of the resulting artifact .zip file.

Doing so, the layer will be part of the SDK content, and
a new "ggp run --orbit" option can deploy these files
to the gamelet.

Test: Run CI.
Bug: http://b/176814898.