Refactor initialization inside the Tracks
Move initialization (name, label, ...) inside the Tracks instead of in
TrackManager. Now GetOrCreate methods are really similar and could be
templatized in the future or we can have an only list of tracks and
easily get from there.

We are using methods SetName and SetLabel in each particular Track
instead of passing them by arguments, but could be changed in the
future.