Process GPU tracepoint events on the ProcessDeferredEvents thread
With this change, all `perf_event_open` events are now processed on a separate
thread than the one that reads from the ring buffers.
This is the last step towards unifying how we process events in `TracerThread`
and towards producing `CaptureEvent`s in order of timestamps.
This also means that individual GPU tracepoint events are now also processed in
order, which solves the potential problem captured by the last test in
`GpuTracepointVisitorTest`: if the tracepoint events are not processed reasonably
in order, some wrong `GpuJob`s could result.
Performance implications of this change are not an issue as these events are
relatively low frequency: small test on this on Trata and Infiltrator with the
next commit.

Note that this currently doesn't work: `dma_fence_signaled` events can come out
of order even in a single ring buffer, which is not supported by
`PerfEventProcessor`/`PerfEventQueue`.
This is going to be solved in the next commit, by adding support for it to
`PerfEventQueue`.