Fix the flickering capture toolbar extension popup
When the user moves the splitter, shrinks down the main window, or works
on a very low resolution screen, the capture toolbar extension toolbar
keeps flickering. The reason is that every time when the process list
updates (i.e., ‘OnProcessListUpdated(...)’ triggered), we reset the icon
of the capture toggle button (i.e., in
`UpdateCaptureStateDependentWidgets()`) and this causes the toolbar to
re-insert every actions.
To fix the above problem, we add the
`UpdateProcessConnectionStateDependentWidgets()` which only updates the
widget affected by the process connection states when the process list
is updated.