Use std::optional rather than address of const ref or address of map element.
This fixes the flaky CaptureEventProcessor test. `FindMatching*` returned a
pointer to a map element. However, this map is not garantueed to be pointer
stable. Thus, we now return std::optional instead.

Also we used the address of operator on a const-ref., which is also
undefined. Again, we are now using std::optional here.

Test: Run CaptureEventProcessorTest a couple of times.