Fix use after deletion.
This fixes the flaky CaptureEventProcessor test. It failed because
"DecrementUnprocessedBeginMarkers" may erase the matching
"GpuQueuSubmission". In some cases, this may be the same as the
current "gpu_queue_submission" to be processed and still accessed
after the call to decrement the marker.

This fixes this by enforcing the call to "DecrementUnprocessedBeginMarkers"
to take place at the very end.

Test: Run CaptureEventProcessor test a couple of times.