Add return type to Future::RegisterContinuation
RegisterContinuation will only register the continuation when the future
hasn't been completed yet. There is no simple race-condition-free way for
the caller to check whether it has been registered or not.

So by returning an enum which tells the caller if the continuation has
been registered or not, the caller can figure out if the future
completed while calling RegisterContinuation.