Add FutureHelpers
This adds two helper functions.

`RegisterContinuationOrCallDirectly` tries to register a continuation.
When that fails due to the future being already completed (often due to
a race condition), the continuation is called directly.

`UnwrapFuture` takes a `Future<Future<T>>` and turns it into a
`Future<T>` by chaining the continuations. The situation often occurs
when chaining tasks with `Future::Then` and the continuation being
asynchronous itself.