Remove font_size parameters
The font_size parameter was bubbling down from main where it was set as
a constant, and then used for both our OpenGl UI and ImGui - where the
resulting on-screen font size was actually completely different.

This change moves the font size used for our capture UI into the
TimeGraphLayout, and replaces the font size for ImGui with a constant
that is defined close to ImGui initialization.