Introduce OrbitApp::LoadModule
While previously `LoadModules` was an implementation that could load
multiple modules at the same time, the new implementation is just a loop
over the function `LoadModule`. That helps to decrease complexity.

There is also still an overload of `LoadModules` which mimics the old
behaviour. That will be removed in a subsequent commit.