Draw overlaps of timers as trinangles.
Especially GPU timers can overlap (due to the pipeline nature).
This change ensures that the text of such timers does not overlap
and that it is clearly visible to the user that these timers do
overlap.

To do so, the overlapping area is drawn as two triangles. Further,
the coloring is slightly adjusted, such that overlapping timers
do not have the same color, and the triangles can be seen.

Test: Need to be done on the feature/vulkan_layer branch.