Update transparent overlays in connection window
The transparent overlays are no longer parented directly underneath the tables,
as this makes them invisible to the E2E tests. Instead, both the tables and the
overlays are parented in a new frame.

This is a bit of a workaround - the nice solution would probably be to provide a
custom table widget that includes the overlay feature, but it would require custom
accessibility information for this widget which is significant amount of work
and not worth the effort.