Remove GCurrentTimeGraph global.
Effectively the global GCurrentTimeGraph is always
app_::capture_window_::time_graph_ (with the fuzzer being the only
exception). This change removes the global and use the value directly
instead.
It also fixes crashes in the fuzzer, which also works with this change
due to the check if capture_window_ exists.

Follow-up: There is an existing bug, that shows the timerange in
the capture window after clearing a capture. We on clearing a capture,
we should not "clear" the time graph, but actually remove it.

Test: Take, clear, load captures.
Bug: http://b/177651082