Fix crash in ScopedStatus' destructor
ScopedStatus used to hold a raw pointer to MainThreadExecutor, which
means it's the creators responsibility to ensure that the
MainThreadExecutor instance stays alive as long as the ScopedStatus
instance is alive.

This doesn't work well with the use-case of ScopedStatus. It is meant to
be used in background threads, which means the creator cannot ensure
that the MainThreadExecutor instance still exists, when the ScopedStatus
object is destructed - at least not without bigger changes to the
codebase.

This commit fixes the problem by making ScopedStatus keep a weak_ptr of
MainThreadExecutor. Whenever it is meant to schedule an action it will
first check whether a lock can still be acquired. If not the action will
be ignored.

Note, it's still the users responsiblity to keep the MainThreadExecutor
alive to make ScopeStatus work properly.

Bug: http://b/177021435