Make LoadSymbol use weak_ptr of MainThreadExecutor
OrbitApp::LoadSymbol (and others) schedule background jobs for
performing expensive tasks.
These background jobs used to capture raw pointers to a
MainThreadExecutor instance, which is accessed at the end of each
background job.

The MainThreadExecutor instance might have been destructed when the
background task reaches its end. That usually happens when the user
closes the main window while the background task is still ongoing.

To avoid the resulting crash, the raw pointer is replaced by a weak_ptr
which allows the background task to check whether the MainThreadExecutor
instance is still alive.

This is a temporary fix, since capturing the main thread executor won't
be necessary anymore with the introduction of Futures.

Bug: http://b/177021435