Add TrySchedule for MainThreadExecutor
It's a helper function that takes a weak_ptr<MainThreadExecutor> and a
function object. It will schedule the function object whenever the
weak_ptr can be locked.

The resulting type is an optional of a Future. So in case the locking
worked it will return the return value of MainThreadExecutor::Schedule,
otherwise a nullopt.