Fix parsing of OpenGL version numbers
I incorrectly assumed that tokenizing the OpenGL version string can
be done with just splitting on ".", but this does not yield correct
results in cases like "3.1 Mesa 20.1.8". Since all major and minor
version numbers are single digits for all OpenGL versions released so
far, we just check the 0th and 2nd digit of the string and consider
them as our major and minor versions. This is wrong for version numbers
with more than one digit, but in our check will only cause issues for
OpenGL 10.x and higher (if we ever see that).

Tested: Ran Orbit manually.
Bug: http://b/179247771